+++ samples/Chart/33_Chart_create_composite.php
@@ -3,7 +3,7 @@
 use PhpOffice\PhpSpreadsheet\Chart\Chart;
 use PhpOffice\PhpSpreadsheet\Chart\DataSeries;
 use PhpOffice\PhpSpreadsheet\Chart\DataSeriesValues;
-use PhpOffice\PhpSpreadsheet\Chart\Legend;
+use PhpOffice\PhpSpreadsheet\Chart\Legend as ChartLegend;
 use PhpOffice\PhpSpreadsheet\Chart\PlotArea;
 use PhpOffice\PhpSpreadsheet\Chart\Title;
 use PhpOffice\PhpSpreadsheet\IOFactory;
@@ -128,7 +128,7 @@ $series3 = new DataSeries(
 // Set the series in the plot area
 $plotArea = new PlotArea(null, [$series1, $series2, $series3]);
 // Set the chart legend
-$legend = new Legend(Legend::POSITION_RIGHT, null, false);
+$legend = new ChartLegend(ChartLegend::POSITION_RIGHT, null, false);
 
 $title = new Title('Average Weather Chart for Crete');