+++ samples/Chart/33_Chart_create_scatter.php
@@ -3,7 +3,7 @@
 use PhpOffice\PhpSpreadsheet\Chart\Chart;
 use PhpOffice\PhpSpreadsheet\Chart\DataSeries;
 use PhpOffice\PhpSpreadsheet\Chart\DataSeriesValues;
-use PhpOffice\PhpSpreadsheet\Chart\Legend;
+use PhpOffice\PhpSpreadsheet\Chart\Legend as ChartLegend;
 use PhpOffice\PhpSpreadsheet\Chart\PlotArea;
 use PhpOffice\PhpSpreadsheet\Chart\Title;
 use PhpOffice\PhpSpreadsheet\IOFactory;
@@ -68,7 +68,7 @@ $series = new DataSeries(
 // Set the series in the plot area
 $plotArea = new PlotArea(null, [$series]);
 // Set the chart legend
-$legend = new Legend(Legend::POSITION_TOPRIGHT, null, false);
+$legend = new ChartLegend(ChartLegend::POSITION_TOPRIGHT, null, false);
 
 $title = new Title('Test Scatter Chart');
 $yAxisLabel = new Title('Value ($k)');