+++ src/PhpSpreadsheet/Calculation/LookupRef/LookupBase.php
@@ -7,6 +7,16 @@ use PhpOffice\PhpSpreadsheet\Calculation\Information\ExcelError;
 
 abstract class LookupBase
 {
+    /**
+     * @param mixed $lookup_array
+     */
+    protected static function validateLookupArray($lookup_array): void
+    {
+        if (!is_array($lookup_array)) {
+            throw new Exception(ExcelError::REF());
+        }
+    }
+
     protected static function validateIndexLookup(array $lookup_array, $index_number): int
     {
         // index_number must be a number greater than or equal to 1